% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_count.R
\name{char_count}
\alias{char_count}
\title{Calculate the Number of Characters in a String}
\usage{
char_count(text)
}
\arguments{
\item{text}{A string or a character vector whose character count is desired.}
}
\value{
An integer representing the number of characters in the input string.
}
\description{
This function takes a string as input
and returns the number of characters in it.
It counts all characters including spaces and punctuation.
}
\examples{
char_count("Today is a sunny day!") # Returns 21
char_count(c("Today", "sunny")) # Returns c(5, 5)
}
